requirements.txt
pandas==2.2.2
openpyxl==3.1.5
duckdb==1.1.1
fastapi==0.114.2
uvicorn==0.30.6
pydantic==2.9.2
scikit-learn==1.5.2
sentence-transformers==3.0.1
hdbscan==0.8.40
umap-learn==0.5.6
python-dotenv==1.0.1
pyyaml==6.0.2